# Conflicts:
#	okapi-core/src/main/java/okapi/web/ModuleWebService.java

+++ okapi-core/src/main/java/okapi/MainVerticle.java
@@ -96,7 +96,7 @@ public class MainVerticle extends AbstractVerticle {
         tenantStore = new TenantStoreMemory();
         break;
       default:
-        System.out.println("FATAL: Unknown storage type '" + storage + "'");
+        logger.fatal("Unknown storage type '" + storage + "'");
         System.exit(1);
     }
     moduleWebService = new ModuleWebService(vertx, ms, moduleStore, timeStampStore );
@@ -110,7 +110,7 @@ public class MainVerticle extends AbstractVerticle {
 
   @Override
   public void start(Future<Void> fut) {
-    if ( mongo.isTransient() ) {
+    if ( mongo != null && mongo.isTransient() ) {
       mongo.dropDatabase(res->{
         if ( res.succeeded()) {
           startListening(fut);
